+++ native/gpu/src/index.d.ts
@@ -2778,7 +2778,7 @@ type PresentMode = "fifo" | "fifoRelaxed" | "immediate" | "mailbox";
 
 interface Renderer {
   getPreferredFormat(): GPUTextureFormat;
-  getAlphaMode(): 'opaque' | 'premultiplied';
+  getAlphaMode(): 'opaque' | 'premultiplied' | 'inherit';
   getCurrentTexture(): GPUTexture;
   getCurrentTextureView(): GPUTextureView;
   swap();
@@ -2839,6 +2839,8 @@ interface WebGPU {
     adapter: GPUAdapter;
     device: GPUDevice;
     renderer: Renderer;
+    requestedAlphaMode: "opaque" | "premultiplied";
+    alphaMode: "opaque" | "premultiplied" | "inherit";
   };
   destroy(context: object);
 }